+++ lib/flor/unit/ganger.rb
@@ -54,6 +54,9 @@ module Flor
          (@unit.loader.tasker(domain, 'ganger', message) ||
           @unit.loader.tasker(domain, 'tasker', message))) ||
         @unit.loader.tasker(domain, tname, message)
+#puts "=" * 80
+#pp tconf
+#puts "=" * 80
 
       fail ArgumentError.new(
         "tasker #{tname.inspect} not found"